Added Create a transaction (issuing-bank)

+++ issuing-bank/src/main/resources/schema.sql
@@ -63,7 +63,7 @@ create table account
 (
     id                    bigint auto_increment primary key,
     account_number        varchar(50),
-    balance               numeric,
+    balance               double,
     currency_id           bigint,
     account_type_id       bigint,
     client_id             bigint,
@@ -97,7 +97,7 @@ create table transaction
 (
     id                  bigint auto_increment primary key,
     transaction_date    date,
-    sum                 numeric,
+    sum                 double,
     transaction_name    varchar(255),
     transaction_type_id bigint,
     account_id          bigint,
